Compatibility fixes for unprivileged `gradle` user (#2)
This change fixes up the changes from #3 that came across as a bit too restrictive, in retrospect. In order to support smooth software lifecycle management, it needs to be possible to install additional SDK components, e.g. transitioning from `build-tools:30.0.1` to `build-tools:30.0.3`. It is not desirable to force a synchronized bump on all current git branches at the same time.